Add trivial compiler
The only backend available for this compiler at the moment is JRuby.

The idea is to start from Beryl ASTs, translate them to JRuby ones
and hand over the resulting trees to JRuby machinery, so it can do
their magic (i.e. more low level compilation passes and optimizations)
and execute the resulting byte code on the JVM like any other JRuby
script.