Set up a basic language runtime
To continue with the idea from the previous commit, a new runner
class is introduced to execute Beryl code.

Beryl ASTs are converted to JRuby ASTs and those are handed over
to a JRuby runtime object for execution.

It works for trivial programs at this point (e.g. put a single
integer in the stack).